feat: Add Application properties components - MEED-7130 - Meeds-io/MIPs#144
