+++ layout-webapp/src/main/resources/locale/portlet/LayoutEditor_en.properties
@@ -45,8 +45,8 @@ layout.top=Top
 layout.right=Right
 layout.bottom=Bottom
 layout.left=Left
-layout.borderColor=Border
-layout.color=Color
+layout.border=Border
+layout.borderColor=Color
 layout.borderSize=Size
 layout.boxShadow=Shadow
 layout.borderRadius=Radius
