+++ layout-webapp/src/main/webapp/vue-app/layout-editor/js/LayoutUtils.js
@@ -60,6 +60,10 @@ export const containerModel = {
   height: null,
   // Used to specify some special CSS Classes to be used on container parent
   cssClass: null,
+  marginTop: null,
+  marginRight: null,
+  marginBottom: null,
+  marginLeft: null,
   borderColor: null,
   borderSize: null,
   boxShadow: null,
@@ -117,6 +121,10 @@ export const applicationModel = {
   theme: null,
   width: null,
   height: null,
+  marginTop: null,
+  marginRight: null,
+  marginBottom: null,
+  marginLeft: null,
   borderColor: null,
   borderSize: null,
   boxShadow: null,
